Cleanups (#213)
* Remove unused "using" statements

* Fix compilation, remove broken unused test method

* Remove redundant return/break/continue

* Remove redundant ToString()

* Remove redundant parenthesis & semicolons

* Remove redundant type parameters

* Remove redundant type specifiers

* Remove redundant qualifiers/initializers

* Remove a few unused parameters

* Adjust namespaces to reflect file layout

* Remove redundant checked/unchecked

* Add readonly where applicable

* Reorder modifiers

* Remove redundant parenthesis

* Remove redundant partial

* Remove redundant casts

* Remove redundant unsafe

* Remove empty initializers

* Minor cleanups

+++ DarkUI/Controls/DarkProgressBar.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
-using System.Linq;
-using System.Text;
 using System.Windows.Forms;
 using DarkUI.Config;
 
@@ -63,7 +60,7 @@ namespace DarkUI.Controls
 
             if (m.Msg == 0x000F)
             {
-                float percentage = (float)(Value - Minimum) / (float)(Maximum - Minimum);
+                float percentage = (Value - Minimum) / (float)(Maximum - Minimum);
 
                 using (Graphics g = CreateGraphics())
                 {
